Move HashWithIndifferentAccess tests to separate file
